refactor dialog and report
