+++ awatif-ui/src/dialog/index.html
@@ -5,66 +5,26 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Awatif UI - Dialog</title>
     <script type="module">
-      import { dialog } from "./dialog";
-      import { sheets } from "awatif-ui/src/sheets/sheets";
-      import { html } from "lit-html";
+      import { getDialog } from "./getDialog";
       import van from "vanjs-core";
 
-      // Define sheets
-      const sheetsObj = new Map();
+      const span = document.createElement("span");
+      span.textContent = "Hello World";
 
-      sheetsObj.set("nodes", {
-        text: "Nodes",
-        fields: [
-          { field: "A", text: "X-coordinate", editable: { type: "float" } },
-          { field: "B", text: "Y-coordinate", editable: { type: "float" } },
-          { field: "C", text: "Z-coordinate", editable: { type: "float" } },
-        ],
-        data: van.state([
-          [0, 0, 0],
-          [1, 2, 3],
-          [3, 4, 1],
-        ]),
-      });
+      const bold = document.createElement("b");
+      bold.textContent = "Mohamed Adil";
 
-      sheetsObj.set("elements", {
-        text: "Elements",
-        fields: [
-          { field: "A", text: "Node 1", editable: { type: "float" } },
-          { field: "B", text: "Node 2", editable: { type: "float" } },
-        ],
-        data: van.state([
-          [0, 0],
-          [1, 2],
-          [3, 4],
-        ]),
-      });
+      const title = van.state("Span");
+      const dialogBody = van.state(span);
 
-      // Create sheets element
-      const sheetsElm = sheets({
-        sheets: sheetsObj,
-        onChange: (e) => console.log(e), // test on change
-      });
+      setTimeout(() => {
+        title.val = "Bold";
+        dialogBody.val = bold;
+      }, 1000);
 
-      // Template for the tables dialog
-      const templateTables = () => {
-        return html`
-          <div>
-            <h2>Tables</h2>
-            <div id="sheets-container">
-              ${sheetsElm} <!-- Render the sheets here -->
-            </div>
-          </div>
-        `;
-      };
+      const dialogElm = getDialog({ title, dialogBody });
 
-      // Create the dialog with the template
-      const dialogObj = dialog({
-        template: templateTables,
-      });
-
-      // Append the dialog to the body
-      document.body.appendChild(dialogObj);
+      document.body.appendChild(dialogElm);
     </script>
   </head>
   <body></body>
