--- awatif-ui/src/viewer/report/style.css
@@ -1,77 +1,0 @@
-.topBar {
-  position: absolute;
-  top: 0.3rem;
-  right: 1rem;
-}
-
-.topBar a {
-  text-decoration: none;
-  color: white;
-}
-
-canvas {
-  position: absolute;
-}
-
-dialog {
-  position: fixed;
-  z-index: 2;
-  overflow: auto;
-  width: calc(100% - 292px - 3rem);
-  height: calc(100% - 3rem);
-  margin: 2rem 1rem 0;
-  padding: 0;
-}
-
-@media only screen and (max-width: 600px) {
-  dialog {
-    width: calc(100% - 2rem);
-    height: calc(70%);
-  }
-}
-
-dialog .close {
-  color: #000000;
-  float: right;
-  margin-top: -5px;
-  font-size: 25px;
-  font-weight: bold;
-  cursor: pointer;
-}
-
-.dialog-header {
-  background: lightgray;
-  padding: 10px;
-  height: 22px;
-}
-
-.dialog-body {
-  padding: 0 15px;
-}
-
-.report-button {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  width: 80px;
-  height: 40px;
-  color: #fefafa;
-  background-color: #2e5368; /* Light red background */
-  border-radius: 10px;
-  font-family: "Arial", sans-serif; /* Set font family */
-  display: flex;
-  font-size: 13px;
-  align-items: center;
-  justify-content: center; /* Center text */
-  cursor: pointer;
-  transition: background-color 0.3s, transform 0.2s;
-}
-
-.report-button:hover {
-  background-color: #333333; /* Darker red on hover */
-  color: #fff;
-}
-
-.report-button:active {
-  transform: scale(0.95);
-}
