+++ awatif-ui/src/viewer/viewer.ts
@@ -19,8 +19,6 @@ import { nodeResults } from "./objects/nodeResults";
 
 import "./styles.css";
 import { drawing, Drawing } from "./drawing/drawing";
-import { TemplateResult } from "lit-html";
-import { report } from "./report/report";
 
 export type SettingsObj = {
   gridSize?: number;
@@ -44,16 +42,11 @@ export function viewer({
   settingsObj,
   objects3D,
   drawingObj,
-  reportObj,
 }: {
   structure?: Structure;
   settingsObj?: SettingsObj;
   objects3D?: State<THREE.Object3D[]>;
   drawingObj?: Drawing;
-  reportObj?: {
-    template: (data: any) => TemplateResult;
-    data: any;
-  };
 }): HTMLDivElement {
   // init
   THREE.Object3D.DEFAULT_UP = new THREE.Vector3(0, 0, 1);
@@ -125,8 +118,6 @@ export function viewer({
       derivedDisplayScale,
       viewerRender,
     });
-
-  if (reportObj) viewerElm.append(report(reportObj));
 
   // on size change
   const resizeObserver = new ResizeObserver((entries) => {
